Restructure locking.
When setting the callback, we need to hold the Future value lock while
setting the callback in the Promise to prevent lost values.

This implies a lock ordering of Future value -> PromiseInner, so restructure
the Promise setting code to take locks in that order. There's a gap
where we could lose a Future after setting the Promise, but that's fine.